requirements.txt
# Citadel Demo — Python dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0

# Data / Stats
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.0.0

# UI
streamlit>=1.28.0
matplotlib>=3.7.0

# Blockchain (Base Sepolia)
web3>=6.0.0
eth-account>=0.9.0
